Move all const context related code to a separate module
